requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Representation Learning & Deep Learning
transformers>=4.30.0
timm>=0.9.0
einops>=0.6.0

# Data Processing
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0
tqdm>=4.65.0

# Experiment Tracking & Visualization
tensorboard>=2.13.0
wandb>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration & Utils
pyyaml>=6.0
hydra-core>=1.3.0
python-dotenv>=1.0.0
omegaconf>=2.3.0

# Testing & Code Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0


